+++ packages/parser-graphql/src/index.ts
@@ -77,6 +77,7 @@ class GraphQLParser {
     schematic: boolean = false,
   ): TypeDefinition {
     const kind = String(type.kind);
+    const optional = true;
 
     // Non-nullable
     if (kind === Kind.NON_NULL_TYPE) {
@@ -90,6 +91,7 @@ class GraphQLParser {
 
       return {
         nullable,
+        optional,
         type: 'array',
         valueType: this.buildAttribute(field, refinedType.type, true, schematic),
       };
@@ -112,6 +114,7 @@ class GraphQLParser {
 
           return {
             nullable,
+            optional,
             type: 'key',
           };
 
@@ -122,6 +125,7 @@ class GraphQLParser {
         case 'Boolean':
           return {
             nullable,
+            optional,
             type: value.toLowerCase(),
           };
 
@@ -130,6 +134,7 @@ class GraphQLParser {
           if (this.enums[value]) {
             return {
               nullable,
+              optional,
               type: 'enum',
               values: this.enums[value],
               valueType: typeof this.enums[value][0],
@@ -140,6 +145,7 @@ class GraphQLParser {
           if (this.shapes[value]) {
             return {
               nullable,
+              optional,
               reference: value,
               type: 'shape',
             };
@@ -149,6 +155,7 @@ class GraphQLParser {
           if (this.unions[value]) {
             return {
               nullable,
+              optional,
               type: 'union',
               valueTypes: this.unions[value],
             };
@@ -158,6 +165,7 @@ class GraphQLParser {
           if (value === this.name) {
             return {
               nullable,
+              optional,
               self: true,
               type: 'reference',
             };
@@ -168,6 +176,7 @@ class GraphQLParser {
 
           return {
             nullable,
+            optional,
             reference: value,
             type: 'reference',
           };